feat(llm): extract Copilot into ace-llm-copilot
Blank-import ace-llm-copilot so init registers copilot. Drop in-tree
CopilotProvider and RegisterLLM("copilot"). Live chat still uses
provider_id=="copilot" with ciphertext APIKey. Decrypt stays on use.

Refs #450
