+++ backend/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/aceobservability/ace-datasource-prometheus v0.0.0-20260906152221-4b460476c203
 	github.com/aceobservability/ace-datasource-victoriametrics v0.0.0-20260906162925-9dc8e06dd775
 	github.com/aceobservability/ace-llm-anthropic v0.0.0-20260906160347-542e9f4acdea
+	github.com/aceobservability/ace-llm-copilot v0.0.0-20260906163911-724e3d6190ac
 	github.com/aceobservability/ace-llm-openai-compat v0.0.0-20260906163134-4077ef660565
 	github.com/alicebob/miniredis/v2 v2.38.0
 	github.com/aws/aws-sdk-go-v2 v1.45.1
