+++ backend/internal/handlers/ai_plugin.go
@@ -2,6 +2,7 @@ package handlers
 
 import (
 	_ "github.com/aceobservability/ace-llm-anthropic"
+	_ "github.com/aceobservability/ace-llm-copilot"
 	_ "github.com/aceobservability/ace-llm-openai-compat"
 
 	"github.com/aceobservability/ace/backend/pkg/llm"
@@ -16,13 +17,3 @@ type (
 )
 
 var ErrUnknownProviderType = llm.ErrUnknownProviderType
-
-func init() {
-	// Copilot's live chat path remains provider_id=="copilot" (user GH token).
-	// Registering the type means a stored provider_type=copilot does not
-	// impersonate OpenAI-compat. LLMConfig.APIKey is ciphertext; the factory
-	// must not receive a decrypted key (CopilotProvider decrypts on use).
-	llm.RegisterLLM("copilot", func(cfg LLMConfig) (AIProvider, error) {
-		return &CopilotProvider{EncryptedGHToken: cfg.APIKey}, nil
-	})
-}
